--- app/src/main/java/com/sq/thed_ck_licker/ecs/systems/CardsSystem.kt
@@ -1,335 +1,0 @@
-package com.sq.thed_ck_licker.ecs.systems
-
-import androidx.compose.runtime.MutableIntState
-import com.sq.thed_ck_licker.R
-import com.sq.thed_ck_licker.ecs.ComponentManager
-import com.sq.thed_ck_licker.ecs.EntityId
-import com.sq.thed_ck_licker.ecs.EntityManager.getPlayerID
-import com.sq.thed_ck_licker.ecs.add
-import com.sq.thed_ck_licker.ecs.components.ActivationCounterComponent
-import com.sq.thed_ck_licker.ecs.components.CardTag
-import com.sq.thed_ck_licker.ecs.components.DescriptionComponent
-import com.sq.thed_ck_licker.ecs.components.DrawDeckComponent
-import com.sq.thed_ck_licker.ecs.components.EffectComponent
-import com.sq.thed_ck_licker.ecs.components.EffectStackComponent
-import com.sq.thed_ck_licker.ecs.components.HealthComponent
-import com.sq.thed_ck_licker.ecs.components.ImageComponent
-import com.sq.thed_ck_licker.ecs.components.MerchantComponent
-import com.sq.thed_ck_licker.ecs.components.NameComponent
-import com.sq.thed_ck_licker.ecs.components.ScoreComponent
-import com.sq.thed_ck_licker.ecs.components.TagsComponent
-import com.sq.thed_ck_licker.ecs.components.activate
-import com.sq.thed_ck_licker.ecs.components.addEntity
-import com.sq.thed_ck_licker.ecs.generateEntity
-import com.sq.thed_ck_licker.ecs.get
-import com.sq.thed_ck_licker.helpers.MyRandom
-import com.sq.thed_ck_licker.helpers.getRandomElement
-import kotlin.math.min
-
-class CardsSystem private constructor(private val componentManager: ComponentManager) {
-
-
-    companion object {
-        val instance: CardsSystem by lazy(LazyThreadSafetyMode.SYNCHRONIZED) {
-            CardsSystem(ComponentManager.componentManager)
-        }
-    }
-
-    /* TODO: multipliers and such can be nicely implemented with:
-*   Make each effect function that takes in the thing  to be multiplied.
-*   Then just pump all things through pipeline full of the multiplier functions.
-*   To get the pipeline just loop all things and collect the multiplier functions.
-*   Something like multiplier system.
-*   If one wants to be wild they can even put three lists as pipeline, one for additions, one for increases and one for multiplications
-*   Thou that would mean order of operation does not matter, which might be wanted.
-*/
-
-
-    // Function to pull a random card from deck
-    fun pullRandomCardFromEntityDeck(entityId: Int): Int {
-        val drawDeck = componentManager.getComponent(entityId, DrawDeckComponent::class)
-        check(drawDeck.drawCardDeck.isNotEmpty()) { "No cards available" }
-        return drawDeck.drawCardDeck.getRandomElement()
-    }
-
-
-    fun activateCard(latestCard: MutableIntState, playerCardCount: MutableIntState) {
-
-        playerCardCount.intValue += 1
-        val latestCardId = latestCard.intValue
-
-        try {
-            (latestCardId get EffectComponent::class).onPlay.invoke(getPlayerID(), latestCardId)
-        } catch (_: Exception) {
-            println("Yeah yeah, we get it, you are so cool there was no effect component")
-        }
-
-        try {
-            (latestCardId get HealthComponent::class).health.floatValue -= 1f
-            println("Health is now ${(latestCardId get HealthComponent::class).health.floatValue}")
-            if ((latestCardId get HealthComponent::class).health.floatValue <= 0) {
-                latestCard.intValue = -1
-            }
-        } catch (_: Exception) {
-            println("Yeah yeah, we get it, you are so cool there was no health component")
-        }
-
-        try {
-            (latestCardId get ActivationCounterComponent::class).activate()
-        } catch (_: Exception) {
-            println("Yeah yeah, we get it, you are so cool there was no actCounter component ")
-        }
-    }
-
-    fun addMerchantCards(amount: Int, merchantId: Int): List<EntityId> {
-        val openMerchant = { id: Int, _: Int ->
-            val target = id get MerchantComponent::class
-            target.merchantId.intValue = merchantId
-        }
-
-        return initCards(
-            null,
-            null,
-            amount,
-            R.drawable.placeholder,
-            "Activate to access shop",
-            "Merchant #$merchantId",
-            listOf(CardTag.CARD),
-            openMerchant,
-            null
-        )
-
-    }
-
-    fun addMaxHpTrapCards(amount: Int = 5): List<EntityId> {
-        val maxHpIt = { playerId: Int, cardEntity: Int ->
-            val playerHp = playerId get HealthComponent::class
-            if (MyRandom.getRandomInt() <= 2) {
-                (cardEntity get HealthComponent::class).health.floatValue -= 99999f
-                playerHp.health.floatValue = (playerHp.health.floatValue.div(2))
-            } else {
-                playerHp.maxHealth.floatValue += 10f
-            }
-        }
-        return initCards(
-            99999f,
-            null,
-            amount,
-            R.drawable.placeholder,
-            "Gain 10 max health on play, might explode",
-            "Max HP Trap Card",
-            listOf(CardTag.CARD),
-            maxHpIt,
-            null
-        )
-    }
-
-
-    fun addBreakingDefaultCards(amount: Int = 7): List<EntityId> {
-        val scoreIt = { playerId: Int, cardEntity: Int ->
-            val target = playerId get ScoreComponent::class
-            val omaScore = cardEntity get ScoreComponent::class
-            target.score.intValue += omaScore.score.intValue
-        }
-        return initCards(
-            10f,
-            100,
-            amount,
-            R.drawable.placeholder,
-            "",
-            "Default Card",
-            listOf(CardTag.CARD),
-            scoreIt,
-            null
-        )
-    }
-
-    private fun initCards(
-        cardHealth: Float?,
-        scoreAmount: Int?,
-        amount: Int,
-        cardImage: Int,
-        description: String,
-        name: String,
-        tags: List<CardTag>,
-        onCardPlay: (Int, Int) -> Unit,
-        onCardDeactivate: (Int, Int) -> Unit = { _, _ -> }
-    ): List<Int> {
-        val cardIds: MutableList<Int> = mutableListOf()
-        repeat(amount) {
-            val cardEntity = generateEntity()
-            cardHealth?.let { health -> cardEntity add HealthComponent(health) }
-            scoreAmount?.let { score -> cardEntity add ScoreComponent(score) }
-            cardEntity add ActivationCounterComponent()
-            cardEntity add ImageComponent(cardImage)
-            cardEntity add EffectComponent(onDeactivate = onCardDeactivate, onPlay = onCardPlay)
-            cardEntity add DescriptionComponent(description)
-            cardEntity add NameComponent(name)
-            cardEntity add TagsComponent(tags)
-            cardIds.add(cardEntity)
-        }
-        return cardIds
-    }
-
-    fun addDeactivationTestCards(amount: Int = 2): List<EntityId> {
-
-        val onActivation = { playerId: Int, cardEntity: Int ->
-            val target = playerId get ScoreComponent::class
-            val riskPoints = cardEntity get ScoreComponent::class
-            val scoreIncrease = riskPoints.score.intValue * 3
-            target.score.intValue += (scoreIncrease)
-            riskPoints.score.intValue = 0
-            println("Now its activated")
-            println("Gave $scoreIncrease points")
-        }
-        val deactivateAction = { playerId: Int, cardEntity: Int ->
-            val target = playerId get HealthComponent::class
-            val riskPoints = cardEntity get ScoreComponent::class
-            riskPoints.score.intValue += 1
-            target.health.floatValue -= riskPoints.score.intValue.toFloat()
-            println("Now its deactivated")
-            println("Risk is rising!")
-            println("Holds ${riskPoints.score.intValue} points")
-
-        }
-
-        for (i in 1..amount) {
-            val cardEntity = generateEntity()
-            cardEntity add ImageComponent()
-            cardEntity add DescriptionComponent("On deactivate you lose health, on activation you gain score * 3")
-            cardEntity add EffectComponent(onDeactivate = deactivateAction, onPlay = onActivation)
-            cardEntity add NameComponent("Deactivation Card #$i")
-            cardEntity add TagsComponent(listOf(CardTag.CARD))
-            cardEntity add riskPoints
-            cardEntity add ActivationCounterComponent()
-        }
-        return initCards(
-            null,
-            0,
-            amount,
-            R.drawable.placeholder,
-            "On deactivate you lose health, on activation you gain score * 3",
-            "Deactivation Card ",
-            listOf(CardTag.CARD),
-            onActivation,
-            { _, _ -> })
-    }
-
-    fun addTrapTestCard(amount: Int = 2): List<EntityId> {
-        val cardIds: MutableList<EntityId> = mutableListOf()
-
-        var scoreLoss = 0
-        var healthLoss = 0
-        val activationComponent = ActivationCounterComponent()
-        val deactivateAction = { id: Int ->
-            val target = id get ScoreComponent::class
-            scoreLoss = ((1 + activationComponent.deactivations.intValue) * 3)
-            target.score.intValue -= (activationComponent.deactivations.intValue * 3)
-            println("Lost $scoreLoss points")
-        }
-        val onActivation = { id: Int ->
-            val target = id get HealthComponent::class
-            healthLoss = activationComponent.activations.intValue + 1
-            target.health.floatValue += activationComponent.activations.intValue
-
-            println("Lost $healthLoss health")
-        }
-        for (i in 1..amount) {
-            val cardEntity = generateEntity()
-            cardIds.add(cardEntity)
-            cardEntity add ImageComponent()
-            cardEntity add DescriptionComponent("On deactivate you lose $scoreLoss score, on activation you lose $healthLoss health")
-            cardEntity add EffectComponent(onDeactivate = deactivateAction, onPlay = onActivation)
-            cardEntity add NameComponent("Trap Card #$i")
-            cardEntity add TagsComponent(listOf(CardTag.CARD))
-        }
-        return cardIds.toList()
-    }
-
-    fun addScoreGainerTestCard(amount: Int = 1): List<EntityId> {
-        val cardIds: MutableList<EntityId> = mutableListOf()
-        val pointsPerCard = 3
-        for (i in 1..amount) {
-            val cardEntity = generateEntity()
-            cardIds.add(cardEntity)
-            val activationComponent = ActivationCounterComponent()
-            cardEntity add ImageComponent()
-            cardEntity add DescriptionComponent("Gain Score gainer on play. \nEvery time you play card you gain $pointsPerCard points")
-            cardEntity add EffectComponent(onPlay = {
-                addPassiveScoreGainerToThePlayer(pointsPerCard + i)
-            })
-            cardEntity add activationComponent
-            cardEntity add NameComponent("Score Gainer Card #$i")
-            cardEntity add TagsComponent(listOf(CardTag.CARD))
-            cardEntity add HealthComponent(1f)
-        }
-        return cardIds.toList()
-    }
-
-
-    private fun addPassiveScoreGainerToThePlayer(amount: Int = 3) {
-
-        val entity = generateEntity()
-
-        val activationComponent = ActivationCounterComponent()
-        val activateAction = { id: Int ->
-            val playerScoreComponent = id get ScoreComponent::class
-            activationComponent.activate()
-            playerScoreComponent.score.intValue += amount
-        }
-        entity add activationComponent
-
-
-        entity add EffectComponent(onTurnStart = activateAction)
-
-        val effStackComp = (getPlayerID() get EffectStackComponent::class)
-        effStackComp addEntity (entity)  // I think i have gone mad from the power
-    }
-
-
-    fun addBeerGogglesTestCard(amount: Int = 1): List<EntityId> {
-        val cardIds: MutableList<EntityId> = mutableListOf()
-        for (i in 1..amount) {
-            val cardEntity = generateEntity()
-            cardIds.add(cardEntity)
-            cardEntity add ImageComponent()
-            cardEntity add DescriptionComponent("Equip Beer Goggles that will heal you bit.")
-            cardEntity add EffectComponent(onPlay = {
-                addLimitedSupplyAutoHealToThePlayer()
-            })
-            cardEntity add ActivationCounterComponent()
-            cardEntity add NameComponent("Beer Goggles Card #$i")
-            cardEntity add TagsComponent(listOf(CardTag.CARD))
-            cardEntity add HealthComponent(1f)
-        }
-        return cardIds.toList()
-    }
-
-    private fun addLimitedSupplyAutoHealToThePlayer(amount: Float = 150f) {
-        val entity = generateEntity()
-        val selfHp = HealthComponent(amount)
-        entity add selfHp
-        val selfActCounter = ActivationCounterComponent()
-        entity add selfActCounter
-
-        entity add EffectComponent(onTurnStart = { id: Int ->
-            val targetHealthComponent = id get HealthComponent::class
-            val targetMaxHp = targetHealthComponent.maxHealth.floatValue
-            val targetHp = targetHealthComponent.health.floatValue
-            if (targetHp < targetMaxHp / 2) {
-                val healAmount = (targetMaxHp * 0.8f) - targetHp
-                val ammm = min(selfHp.health.floatValue, healAmount)
-                selfHp.health.floatValue -= ammm
-                targetHealthComponent.health.floatValue += ammm
-            }
-            println("My name is Beer Goggles")
-            println("I am now at ${selfHp.health.floatValue} health \nand have been activated ${selfActCounter.activations.intValue} times")
-        })
-
-        val effStackComp = (getPlayerID() get EffectStackComponent::class)
-        effStackComp addEntity (entity)
-
-
-    }
-
-}
